+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - [4481](https://github.com/vegaprotocol/vega/pull/4481) - Add an example client application that uses the null-blockchain
 - [4503](https://github.com/vegaprotocol/vega/pull/4503) - Limit the number of epochs for which we keep delegations in memory in core api
 - [4514](https://github.com/vegaprotocol/vega/pull/4514) - Add network limits service and events
-- [4489](https://github.com/vegaprotocol/vega/pull/4489) - Check when `OnGenesis` that the has of contract bytecode is as expected
+- [4489](https://github.com/vegaprotocol/vega/pull/4489) - Check when `OnGenesis` that the hash of contract bytecode is as expected
 - [4516](https://github.com/vegaprotocol/vega/pull/4516) - Add a command to cleanup all vega node state
 
 ### 🐛 Fixes